+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 3.7.0 (2025-12-01)
+
+## Features
+
+* **Partitions**: Add comprehensive support for partitions ([#213](https://github.com/matteobortolazzo/couchdb-net/pull/213))
+* **Get/SetRevisionLimit**: Added method to get the revision limit of a database ([#202](https://github.com/matteobortolazzo/couchdb-net/pull/202))
+* **ThrowOnQueryWarning**: Added option to throw exception on query warnings ([#205](https://github.com/matteobortolazzo/couchdb-net/pull/205))
+
+## Bugs
+
+* **IncludeExecutionStats**: Fixed deserialization exception ([#204](https://github.com/matteobortolazzo/couchdb-net/pull/204))
+* **Local Docs**: Fixed ID encoding ([#206](https://github.com/matteobortolazzo/couchdb-net/pull/206))
+
 # 3.6.1 (2024-04-23)
 
 ## Bugs